replace v1 cache with v2 cache
