+++ packages/lix-sdk/src/database/schema.ts
@@ -35,7 +35,7 @@ import {
 import type { EntityViews } from "../entity-views/entity-view-builder.js";
 import type { ToKysely } from "../entity-views/types.js";
 import type { InternalStateCacheTable } from "../state/cache/schema.js";
-import type { InternalStateCacheV2Table } from "../state/cache-v2/schema.js";
+import type { InternalStateCacheV2Table } from "../state/cache/schema.js";
 import type { InternalResolvedStateAllView } from "../state/resolved-state-view.js";
 import type { InternalStateAllUntrackedTable } from "../state/untracked/schema.js";
 import type { InternalFileDataCacheTable } from "../file/cache/schema.js";
